refactor(constants): split monolithic constants.js into domain files
A single constants.js mixing validation rules, UI labels and CSS class
names violates SRP at the module level and grows unboundedly with the app.

- Created src/constants/validation.js for business logic limits
- Created src/constants/ui.js for labels, placeholders and CSS classes
- Created src/constants/index.js re-exporting both for backward
  compatibility — existing imports from 'constants' continue to work
- Updated all consumers to import from the precise domain subfile
- Removed src/constants.js
