+++ src/components/Header/Header.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import { LOGIN_LABEL } from '../../constants';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { logoutUser } from '../../store/user/thunk';
+import { LOGIN_LABEL } from '../../constants/ui';
 import Logo from './components/Logo/Logo';
 import Button from '../../common/Button/Button';
-
 import './Header.css';
 
 function Header() {
