+++ src/constants/ui.js
@@ -1,8 +1,3 @@
-export const MIN_PASSWORD_LENGTH = 6;
-export const MIN_COURSE_TITLE_LENGTH = 2;
-export const MIN_COURSE_DESCRIPTION_LENGTH = 2;
-export const MAX_AUTHORS_DISPLAY_LENGTH = 30;
-
 export const INPUT_PLACEHOLDER = 'Input text';
 export const ADD_NEW_COURSE_LABEL = 'ADD NEW COURSE';
 export const BACK_TO_COURSES_LABEL = 'BACK TO COURSES';
